volume is now int

+++ src/Panacea.Modularity.ScreenCast/IScreenCastPlayer.cs
@@ -5,6 +5,6 @@ namespace Panacea.Modularity.ScreenCast
     public interface IScreenCastPlayer : IMediaPlayer
     {
         void Play(string url, MediaItem media);
-        void SetVolume(float value);
+        void SetVolume(int value);
     }
 }
